test: consolidate Kling official contract tests
